requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
beautifulsoup4==4.9.3    
cachetools==5.3.3        
certifi==2024.2.2        
charset-normalizer==3.3.2
click==7.1.2
fastapi==0.110.0
google-auth==2.29.0      
google-auth-oauthlib==1.2.0
gspread==6.0.2
h11==0.14.0
httplib2==0.22.0
idna==3.7
jugaad-data==0.26
lxml==5.2.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
pandas==1.5.2
protobuf==5.26.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2022.7.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
StrEnum==0.4.15
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.1